ble-sensor-receive-daemon/requirements.txt
bluepy ~= 1.3.0
paho-mqtt ~= 1.5.1
pyyaml ~= 5.3.1
pycrypto ~= 2.6.1
